Apply PerfumeSummaryDTO
